pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Advanced Document Q&A RAG System with Multi-Agent Architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.129.0",
    "fastmcp>=2.14.5",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-milvus>=0.3.3",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "langchain-experimental>=0.0.49",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "markdown>=3.10.2",
    "mcp>=1.26.0",
    "numpy>=2.2.6",
    "openai>=2.20.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymilvus>=2.6.9",
    "pypdf>=6.7.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sse-starlette>=3.2.0",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
